fix(navigation): Correct DotRecast poly refs, area types, obstacle wiring, agent/grid pathing (#1168-#1170,#1172-#1174)
Navigation correctness cluster:

- #1169: Widen NavPoint.PolygonId (and INavigationMesh.GetPolygonVertices)
  from uint to long and stop truncating DotRecast poly refs. DotRecast packs
  salt/tile bits above bit 32, so the uint cast broke GetTileAndPolyByRef's
  salt check and returned PolygonIds failed round-trip (0 vertices).
- #1170: Remap Recast's RC_WALKABLE_AREA (63) sentinel to NavAreaType.Walkable
  (0) on walkable ground polys during build, so area type is in the 0-31 range
  and SetAreaCost can influence terrain cost.
- #1168: Consult NavMeshObstacleManager during query-filter construction in
  DotRecastProvider (via a wrapping exclusion filter) and wire the manager into
  the provider on plugin install, so manager-added obstacles actually exclude
  polygons from pathfinding.
- #1172: Check DtNavMesh.AddTile status in BuildTiled and throw a diagnostic on
  failure instead of silently dropping tiles past the 16384-tile cap.
- #1173: Clear NavMeshAgent.PathPending when a path request cannot be enqueued
  (no provider or no Transform3D), so an agent is not left stuck computing.
- #1174: Scan one extra ring in GridNavigationProvider.FindNearestPoint before
  committing, since a straight-line cell in ring r+1 can be closer than a
  diagonal cell in ring r.

Closes #1168
Closes #1169
Closes #1170
Closes #1172
Closes #1173
Closes #1174

+++ src/KeenEyes.Navigation.Abstractions/INavigationMesh.cs
@@ -103,5 +103,5 @@ public interface INavigationMesh
     /// </summary>
     /// <param name="polygonId">The polygon identifier.</param>
     /// <returns>The vertices of the polygon, or an empty span if invalid.</returns>
-    ReadOnlySpan<Vector3> GetPolygonVertices(uint polygonId);
+    ReadOnlySpan<Vector3> GetPolygonVertices(long polygonId);
 }
